move the flag to codegen and remove scattering to HAL
