+++ compiler/src/iree/compiler/Codegen/ExternalInterfaces/CMakeLists.txt
@@ -43,7 +43,6 @@ iree_cc_library(
     iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::LinalgExt::Utils
     iree::compiler::Dialect::TensorExt::IR
-    iree::compiler::Utils
   PUBLIC
 )
