+++ compiler/src/iree/compiler/Utils/BUILD.bazel
@@ -85,6 +85,5 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:Support",
         "@llvm-project//mlir:TransformUtils",
         "@llvm-project//mlir:Transforms",
-        "@llvm-project//mlir:VectorDialect",
     ],
 )
