+++ compiler/src/iree/compiler/Utils/EquivalenceUtils.h
@@ -8,7 +8,6 @@
 #define IREE_COMPILER_UTILS_EQUIVALENCEUTILS_H_
 
 #include "llvm/ADT/SetVector.h"
-#include "mlir/IR/Builders.h"
 #include "mlir/IR/Operation.h"
 
 namespace mlir {
@@ -17,19 +16,6 @@ class IRMapping;
 } // namespace mlir
 
 namespace mlir::iree_compiler {
-
-// Returns the runtime value of vscale specified by the user. This is not meant
-// to be used for codegen and is meant to circumvent the current limitation on
-// host-side querying of this value at runtime. This is temporary until #21317
-// is resolved.
-unsigned getUserVscaleValue();
-
-// If the given operation is a `vector.vscale`, generates the constant value for
-// it, maps the original op in bvm and adds the operation to the vector of
-// vscale ops. Otherwise, it's a noop. This is temporary until #21317 is
-// resolved.
-void mapVscaleOpToConstant(Operation &op, OpBuilder &builder, IRMapping &bvm,
-                           SmallVector<Operation *> &vscaleOps);
 
 // Recursively compares two regions for structural equivalence.
 //
